+++ R/rrs.R
@@ -6,7 +6,10 @@
 #' @param alpha Numeric. The two-tailed alpha level for significance threshold (default is 0.05).
 #' @param p_correct Character string specifying whether to apply a correction for multiple comparisons including a Bonferroni correction \code{p_correct = "uncorrelated"} or a correlated Bonferroni correction \code{p_correct = "correlated"}. If \code{p_correct = "none"} then no correction is applied. 
 #' @param nbc Optional. An integer for the number of bins when \code{p_correct = "correlated"}. Similar to \code{nbclass} argument in \code{\link[pgirmess]{correlog}}. The default is the average number of gridded knots in one-dimension (i.e., x-axis). 
-#' @param doplot Logical. If \code{TRUE}, the output includes basic data visualizations.
+#' @param plot_gate Logical. If \code{TRUE}, the output includes basic data visualization.
+#' @param save_gate Logical. If \code{TRUE}, the output saves the visualization as a separate PNG file.
+#' @param name_gate Optional, character. The filename of the visualization. The default is "gate".
+#' @param path_gate Optional, character. The path of the visualization. The default is the current working directory.
 #' @param rcols Character string of length three (3) specifying the colors for: 1) group A (numerator), 2) neither, and 3) group B (denominator) designations. The defaults are \code{c("#FF0000", "#cccccc", "#0000FF")} or \code{c("red", "grey80", "blue")}.
 #' @param lower_lrr Optional, numeric. Lower cut-off value for the log relative risk value in the color key (typically a negative value). The default is no limit and the color key will include the minimum value of the log relative risk surface. 
 #' @param upper_lrr Optional, numeric. Upper cut-off value for the log relative risk value in the color key (typically a positive value). The default is no limit and the color key will include the maximum value of the log relative risk surface.
@@ -15,7 +18,7 @@
 #' @param verbose Logical. If \code{TRUE} will print function progress during execution. If \code{FALSE} (the default), will not print.
 #' @param ... Arguments passed to \code{\link[sparr]{risk}} to select bandwidth, edge correction, and resolution.
 #'
-#' @details This function estimates a relative risk surface and computes the asymptotic p-value surface for a single gate and single condition using the \code{\link[sparr]{risk}} function. Bandwidth is fixed across both layers (numerator and denominator spatial densities). Basic visualization is available if \code{doplot = TRUE}. 
+#' @details This function estimates a relative risk surface and computes the asymptotic p-value surface for a single gate and single condition using the \code{\link[sparr]{risk}} function. Bandwidth is fixed across both layers (numerator and denominator spatial densities). Basic visualization is available if \code{plot_gate = TRUE}. 
 #' 
 #' Provides functionality for a correction for multiple testing. If \code{p_correct = "uncorrelated"}, then a conventional Bonferroni correction is calculated by dividing the \code{alpha} level by the number of gridded knots across the estimated surface. The default in the \code{\link[sparr]{risk}} function is a resolution of 128 x 128 or n = 16,384 knots and a custom resolution can be specified using the \code{resolution} argument within the \code{\link[sparr]{risk}} function. If \code{p_correct = "correlated"} (NOTE: May take a considerable amount of computation resources and time), then a Bonferroni correction that takes into account the spatial correlation of the surface is calculated within the internal \code{pval_correct} function. The \code{alpha} level is divided by the minimum number of knots that are not spatially correlated. The minimum number of knots that are not spatially correlated is computed by counting the knots that are a distance apart that exceeds the minimum distance of non-significant spatial correlation based on a correlogram using the \code{\link[pgirmess]{correlog}} function. If \code{p_correct = "none"}, then the function does not account for multiple testing and uses the uncorrected \code{alpha} level. See the internal \code{pval_correct} function documentation for more details.
 #' 
@@ -34,7 +37,7 @@
 #'
 #' @importFrom fields image.plot
 #' @importFrom graphics close.screen par screen split.screen 
-#' @importFrom grDevices chull
+#' @importFrom grDevices chull dev.off png
 #' @importFrom raster extent values
 #' @importFrom spatstat owin ppp
 #' @importFrom sparr OS risk
@@ -48,7 +51,10 @@ rrs <- function(dat,
                 alpha = 0.05, 
                 p_correct = "none",
                 nbc = NULL,
-                doplot = FALSE, 
+                plot_gate = FALSE,
+                save_gate = FALSE,
+                name_gate = NULL,
+                path_gate = NULL,
                 rcols = c("#FF0000", "#cccccc", "#0000FF"),
                 lower_lrr = NULL,
                 upper_lrr = NULL,
@@ -128,7 +134,7 @@ rrs <- function(dat,
   } 
   if (p_correct == "uncorrelated") { out$alpha <- alpha / prod(out$rr$dim) }
   
-  if (doplot == TRUE) {
+  if (plot_gate == TRUE) {
     # Graphics
     op <- graphics::par(no.readonly = TRUE)
     on.exit(graphics::par(op))
@@ -157,6 +163,13 @@ rrs <- function(dat,
       brp <- c(1, 1.67, 2.33, 3)
       atp <- c(1.33, 2, 2.67)
       labp <- c("numerator", "insignificant", "denominator")
+    }
+    
+    # If save plot as a PNG file
+    if (save_gate == TRUE) { 
+      if (is.null(name_gate)) { name_gate <- "gate" } # capture name 
+      gate_file <- paste(path_gate, name_gate, ".png", sep = "") # set filename for image
+      grDevices::png(filename = gate_file, width = 800, height = 600) # open graphics device
     }
     
     graphics::par(pty = "s", bg = "white")
@@ -202,6 +215,7 @@ rrs <- function(dat,
                                     sep = ""))
     graphics::close.screen(all = TRUE) # exit split-screen mode
   }
+  if (save_gate == TRUE) { grDevices::dev.off() } # close graphics device
   
   return(out)
 }